io namespace is not shared publicly, due to public opinion about the io namespace.
Due to how much people complain about hacked clients, and even hacked-client users having a preference for hacked-clients being less possible.  I have decided that the 'io' namespace is going to remain private from this point on until further notice.  It is possible that I will never release the 'io' namespace to the public.  Seeing the struggles that minecraft has due to the 'io' being public, at least to the extent were there are hacked clients, this game-engine might never have its 'io' namespace revealed, and the public may only have access to a compiled version of it.